# apps/api-go/internal/server/openapi.json
{
  "openapi": "3.0.3",
  "info": {"title": "Liquor Store Security API", "description": "Go backend contract for the store owner security dashboard.", "version": "1.0.0"},
  "servers": [{"url": "http://localhost:3000"}],
...
  "paths": {
...
    "/api/v1/internal/ai/alerts": {"post":{"tags":["internal"],"summary":"Ingest an AI-generated alert and enqueue matching notifications","description":"Service-to-service endpoint enabled only when AI_INGEST_TOKEN is configured. sourceEventId is idempotent. The alert, evidence and matching delivery outbox rows commit in one transaction.","security":[{"aiIngestAuth":[]}],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/AIAlertIngestRequest"}}}},"responses":{"201":{"description":"Alert, evidence and delivery routes created"},"200":{"description":"Existing sourceEventId returned without duplicate rows"},"400":{"description":"Invalid alert or evidence contract"},"401":{"description":"Invalid AI service token"},"404":{"description":"Store, camera or zone not found"},"409":{"description":"Conflicting source event or evidence key"}}}},
...
    "/api/v1/stores/{storeId}/alerts": {"parameters":[{"$ref":"#/components/parameters/StoreId"}],"get":{"tags":["alerts"],"summary":"List and filter alerts","security":[{"bearerAuth":[]}],"parameters":[{"name":"status","in":"query","schema":{"type":"string"}},{"name":"severity","in":"query","schema":{"type":"string"}},{"name":"type","in":"query","schema":{"type":"string"}},{"name":"cameraId","in":"query","schema":{"type":"string"}},{"name":"cursor","in":"query","schema":{"type":"string","format":"date-time"}},{"name":"limit","in":"query","schema":{"type":"integer","default":30,"maximum":100}}],"responses":{"200":{"description":"Alert page"}}}},
    "/api/v1/stores/{storeId}/alerts/{alertId}": {"parameters":[{"$ref":"#/components/parameters/StoreId"},{"$ref":"#/components/parameters/AlertId"}],"get":{"tags":["alerts"],"summary":"Get alert detail and evidence","security":[{"bearerAuth":[]}],"responses":{"200":{"description":"Alert detail"}}}},
...
  },
  "components": {
    "securitySchemes":{"bearerAuth":{"type":"http","scheme":"bearer","bearerFormat":"JWT"},"aiIngestAuth":{"type":"http","scheme":"bearer","bearerFormat":"opaque service token"}},
    "parameters":{
      "StoreId":{"name":"storeId","in":"path","required":true,"schema":{"type":"string","format":"uuid"}},
...
      "AlertId":{"name":"alertId","in":"path","required":true,"schema":{"type":"string","format":"uuid"}},
...
    },
...
    "schemas":{
...
      "AIAlertIngestRequest":{"type":"object","required":["sourceEventId","storeId","cameraId","type","severity","detectedAt","evidence"],"properties":{"sourceEventId":{"type":"string","minLength":1,"maxLength":200},"correlationId":{"type":"string","maxLength":200},"storeId":{"type":"string","format":"uuid"},"cameraId":{"type":"string","format":"uuid"},"zoneId":{"type":"string","format":"uuid"},"type":{"type":"string","enum":["CASH_DRAWER_WITHOUT_CUSTOMER","SUSPICIOUS_CASH_HANDLING","POS_VOID_OR_REFUND","UNAUTHORIZED_STOCKROOM_ACCESS","HIGH_VALUE_ZONE_DWELL","WEAPON_DETECTED","SUSPICIOUS_PRODUCT_CONCEALMENT"]},"severity":{"type":"string","enum":["LOW","MEDIUM","HIGH","CRITICAL"]},"subjectPersonCategory":{"type":"string","enum":["EMPLOYEE","CUSTOMER","UNKNOWN"],"default":"UNKNOWN"},"confidence":{"type":"number","minimum":0,"maximum":1},"detectedAt":{"type":"string","format":"date-time"},"observedStartAt":{"type":"string","format":"date-time"},"observedEndAt":{"type":"string","format":"date-time"},"metadata":{"type":"object","additionalProperties":true,"maxProperties":100},"evidence":{"type":"array","minItems":1,"maxItems":5,"items":{"$ref":"#/components/schemas/AIAlertEvidence"}}}},
      "AIAlertEvidence":{"type":"object","required":["storageKey","mimeType","durationSeconds","startsAt","endsAt"],"properties":{"storageKey":{"type":"string","minLength":1,"maxLength":500,"description":"Relative object key only; URLs and traversal segments are rejected"},"mimeType":{"type":"string","enum":["video/mp4","video/webm"]},"durationSeconds":{"type":"integer","minimum":1,"maximum":600},"startsAt":{"type":"string","format":"date-time"},"endsAt":{"type":"string","format":"date-time"}}},
...
    }
  }
}